refactor: move components to their own directories and rename Expenses to ExpensesPage
