+++ src/modules/auth/components/AuthorizationForm/AuthorizationForm.tsx
@@ -1,10 +1,10 @@
 import { InputField, Button, T } from '@admiral-ds/react-ui';
-import { InputsStyled } from '../../shared/styled';
-import { RoutesEnum } from '../../../app/routes/RoutesEnum';
+import { InputsStyled } from '../../../shared/styled';
+import { RoutesEnum } from '../../../../app/routes/RoutesEnum';
 import { Link } from 'react-router-dom';
 import { Controller } from 'react-hook-form';
 import { FormStyled, SpinnerLoader } from './AuthorizationForm.styled';
-import { useAuthorizationForm } from '../hooks/useAuthorizationForm';
+import { useAuthorizationForm } from '../../hooks/useAuthorizationForm';
 export const AuthorizationForm = () => {
     const { control, onSubmit, isLoading } = useAuthorizationForm();
