+++ src/modules/auth/components/RegistrationForm/RegistrationForm.styled.ts
