+++ src/modules/auth/components/RegistrationForm/RegistrationForm.tsx
@@ -1,10 +1,10 @@
 import { InputField, Button, SelectField, Option, T } from '@admiral-ds/react-ui';
 import { Link } from 'react-router-dom';
 import { Controller } from 'react-hook-form';
-import { InputsStyled } from '../../shared/styled';
-import { RoutesEnum } from '../../../app/routes/RoutesEnum';
+import { InputsStyled } from '../../../shared/styled';
+import { RoutesEnum } from '../../../../app/routes/RoutesEnum';
 import { FormStyled } from './RegistrationForm.styled';
-import { useRegistrationForm } from '../hooks/useRegistrationForm';
+import { useRegistrationForm } from '../../hooks/useRegistrationForm';
 
 const selectOptions = [
     { value: 'admin', label: 'Да' },
